+++ plugins/basesensors/baselaser.h
@@ -32,7 +32,7 @@ protected:
                     return PE_Support;
                 return PE_Ignore;
             }
-            if( name != "sensor" && name != "minangle" && name != "maxangle" && name != "maxrange" && name != "minrange" && name != "scantime" && name != "color" && name != "resolution" && name != "time_scan" && name != "time_increment" ) {
+            if( name != "sensor" && name != "minangle" && name != "maxangle" && name != "maxrange" && name != "minrange" && name != "scantime" && name != "color" && name != "resolution" && name != "time_scan" && name != "time_increment" && name != "power") {
                 return PE_Pass;
             }
             ss.str("");
@@ -48,6 +48,9 @@ protected:
             }
             else if( name == "sensor" ) {
                 return true;
+            }
+            else if( name == "power" ) {
+                ss >> _psensor->_bPower;
             }
             else if( name == "minangle" ) {
                 ss >> _psensor->_pgeom->min_angle[0];
@@ -198,10 +201,10 @@ public:
             {
                 // Lock the data mutex and fill with the range data (get all in one timestep)
                 boost::mutex::scoped_lock lock(_mutexdata);
-                _pdata->t = GetTransform();
+                //_pdata->t = GetTransform();
                 _pdata->__stamp = GetEnv()->GetSimulationTime();
-        
                 t = GetLaserPlaneTransform();
+                _pdata->positions.at(0) = t.trans;
                 size_t index = 0;
                 for(dReal frotangle = _pgeom->min_angle[0]; frotangle <= _pgeom->max_angle[0]; frotangle += _pgeom->resolution[0], ++index) {
                     if( index >= _pdata->ranges.size() ) {
@@ -242,8 +245,9 @@ public:
                     boost::mutex::scoped_lock lock(_mutexdata);
                     N = (int)_pdata->ranges.size();
                     vpoints.resize(N+1);
-                    for(int i = 0; i < N; ++i)
+                    for(int i = 0; i < N; ++i) {
                         vpoints[i] = _pdata->ranges[i] + t.trans;
+                    }
                     vpoints[N] = t.trans;
                 }
 
@@ -350,7 +354,7 @@ protected:
     {
         boost::mutex::scoped_lock lock(_mutexdata);
         int N = (int)( (_pgeom->max_angle[0]-_pgeom->min_angle[0])/_pgeom->resolution[0] + 0.5f)+1;
-        _pdata->positions.clear();
+        _pdata->positions.resize(1);
         _pdata->ranges.resize(N);
         _pdata->intensity.resize(N);
         _databodyids.resize(N);
@@ -498,8 +502,9 @@ public:
     {
         if( _bPower ) {
             _fCurAngle += _fGeomSpinSpeed*fTimeElapsed;
-            if( _fCurAngle > 2*PI )
+            if( _fCurAngle > 2*PI ) {
                 _fCurAngle -= 2*PI;
+            }
             if( _fTimeToScan <= fTimeElapsed ) {
                 // have to update
                 SetTransform(_trans);